--- packages/client/src/pages/500/500.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import { ErrorComponent } from '../../components/error/ErrorComponent'
-
-const ServerErrorPage: React.FC = () => {
-  return (
-    <ErrorComponent
-      type={'500'}
-      message={'Что-то Пошло Не Так... Уже Чиним!'}
-    />
-  )
-}
-
-export default ServerErrorPage
